[ui] Refactor filter creations to separate file per filter
