+++ app/src/app/search/filters/region-filter.ts
@@ -1,0 +1,31 @@
+import { Tables } from "@/lib/database.types";
+import { createURLParamsResolver } from "@/app/search/filters/url-params-resolver";
+
+export const PHYSICAL_REGION_PATH: SearchFilterName = "physical_region_path";
+
+export const createRegionFilter = (
+  params: URLSearchParams,
+  regions: Tables<"region_used">[]
+): SearchFilter => {
+  const [region] = createURLParamsResolver(params)(PHYSICAL_REGION_PATH);
+  return {
+    type: "radio",
+    name: PHYSICAL_REGION_PATH,
+    label: "Region",
+    operator: "cd",
+    options: [
+      {
+        label: "Missing",
+        value: null,
+        humanReadableValue: "Missing",
+        className: "bg-orange-200",
+      },
+      ...regions.map(({ code, path, name }) => ({
+        label: `${code} ${name}`,
+        value: path as string,
+        humanReadableValue: `${code} ${name}`,
+      })),
+    ],
+    selected: region ? [region === "null" ? null : region] : [],
+  };
+};
