+++ app/src/app/search/filters/sector-filter.ts
@@ -1,0 +1,24 @@
+import { Tables } from "@/lib/database.types";
+import { createURLParamsResolver } from "@/app/search/filters/url-params-resolver";
+
+export const SECTOR_CODE: SearchFilterName = "sector_code";
+
+export const createSectorFilter = (
+  params: URLSearchParams,
+  sectors: Tables<"sector">[]
+): SearchFilter => {
+  const [sectorCode] = createURLParamsResolver(params)(SECTOR_CODE);
+  return {
+    type: "options",
+    name: SECTOR_CODE,
+    label: "Sector",
+    operator: "in",
+    options: [
+      ...sectors.map(({ code, name }) => ({
+        label: `${code} ${name}`,
+        value: code,
+      })),
+    ],
+    selected: sectorCode?.split(",") ?? [],
+  };
+};
