+++ app/src/app/search/filters/unit-type-search-filter.ts
@@ -1,0 +1,40 @@
+import { createURLParamsResolver } from "@/app/search/filters/url-params-resolver";
+
+export const UNIT_TYPE: SearchFilterName = "unit_type";
+
+export const createUnitTypeSearchFilter = (
+  params: URLSearchParams
+): SearchFilter => {
+  const [unitType] = createURLParamsResolver(params)(UNIT_TYPE);
+  return {
+    type: "options",
+    label: "Type",
+    name: UNIT_TYPE,
+    operator: "in",
+    options: [
+      {
+        label: "Legal Unit",
+        value: "legal_unit",
+        humanReadableValue: "Legal Unit",
+        className: "bg-legal_unit-100",
+      },
+      {
+        label: "Establishment",
+        value: "establishment",
+        humanReadableValue: "Establishment",
+        className: "bg-establishment-100",
+      },
+      {
+        label: "Enterprise",
+        value: "enterprise",
+        humanReadableValue: "Enterprise",
+        className: "bg-enterprise-100",
+      },
+    ],
+    selected: unitType?.split(",") ?? [
+      "legal_unit",
+      "establishment",
+      "enterprise",
+    ],
+  };
+};
